Integrating some fixes from the main branch. Added ATH1 busy mode, ATNx no-listener feature. Moved IRC client to ESP32 only -- the above just wouldn't fit with it.
git-svn-id: svn://192.168.1.10/public/Zimodem-371@22675 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ zimodem/wifiservernode.ino
@@ -55,7 +55,7 @@ WiFiServerNode::WiFiServerNode(int newport, int flagsBitmap)
   id=++WiFiNextClientId;
   port=newport;
   this->flagsBitmap = flagsBitmap;
-  server = new WiFiServer(newport);
+  server = new WiFiServer((uint16_t)newport);
   //BZ:server->setNoDelay(DEFAULT_NO_DELAY);
   server->begin();
   if(servs==null)
